fix OD radio page tests
